Server room improvements, spark sfx
