--- Assets/GlassMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ddbf3b8a0baa10d4e8becc4ec5c1642d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
