+++ Assets/Prefabs/GrayboxPrefabs/AutomaticDoor Variant (1).prefab
@@ -86,6 +86,8 @@ Transform:
   - {fileID: 4377307417316718911}
   - {fileID: 6406000049122063860}
   - {fileID: 6027780048456882213}
+  - {fileID: 32225944439873963}
+  - {fileID: 1133995261089164675}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: -0.001, y: 90, z: -360}
 --- !u!114 &1036688137448871242
@@ -117,7 +119,9 @@ MonoBehaviour:
   showSparks: 0
   audioSource: {fileID: 5726619166760037528}
   audioManager: {fileID: 0}
-  sparks: []
+  sparks:
+  - {fileID: 8311989325657523217}
+  - {fileID: 8976977791794153529}
 --- !u!1 &1540660211456931277
 GameObject:
   m_ObjectHideFlags: 0
@@ -8826,3 +8830,149 @@ MonoBehaviour:
   - 0
   applyHRTFToPathing: 0
   pathingMixLevel: 1
+--- !u!1001 &3576037090760361517
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 4933824390585207861}
+    m_Modifications:
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.262
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.004
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -1.566
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.9901833
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 4.5474724e-13
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0.13977483
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 4.5474724e-13
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 16.07
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8251244556969653171, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_Name
+      value: PS_Sparks
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+--- !u!4 &32225944439873963 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+  m_PrefabInstance: {fileID: 3576037090760361517}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &8311989325657523217 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 4826377248061997628, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+  m_PrefabInstance: {fileID: 3576037090760361517}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 650081bcc202e2346b32c5edcbf16563, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!1001 &4498565189023910405
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 4933824390585207861}
+    m_Modifications:
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0.274
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.01
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -1.557
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.9932314
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0.11615225
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -9.094949e-13
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: -13.34
+      objectReference: {fileID: 0}
+    - target: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8251244556969653171, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+      propertyPath: m_Name
+      value: PS_Sparks (1)
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+--- !u!4 &1133995261089164675 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+  m_PrefabInstance: {fileID: 4498565189023910405}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &8976977791794153529 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 4826377248061997628, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+  m_PrefabInstance: {fileID: 4498565189023910405}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 650081bcc202e2346b32c5edcbf16563, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
