+++ Assets/Prefabs/GrayboxPrefabs/CylinderDoor.prefab
@@ -1620,8 +1620,8 @@ MonoBehaviour:
   audioSource: {fileID: 3397945899310577105}
   audioManager: {fileID: 0}
   sparks:
-  - {fileID: 6007825717453173431}
-  - {fileID: 4420007780780473013}
+  - {fileID: 73971181077900484}
+  - {fileID: 8012427542703301830}
 --- !u!1 &7173120277529721626
 GameObject:
   m_ObjectHideFlags: 0
@@ -8326,11 +8326,17 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
   m_PrefabInstance: {fileID: 3299837722224600826}
   m_PrefabAsset: {fileID: 0}
---- !u!198 &4420007780780473013 stripped
-ParticleSystem:
-  m_CorrespondingSourceObject: {fileID: 1196942945253612623, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+--- !u!114 &8012427542703301830 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 4826377248061997628, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
   m_PrefabInstance: {fileID: 3299837722224600826}
   m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 650081bcc202e2346b32c5edcbf16563, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &3492683247144627608
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -8466,11 +8472,17 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
---- !u!198 &6007825717453173431 stripped
-ParticleSystem:
-  m_CorrespondingSourceObject: {fileID: 1196942945253612623, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
+--- !u!114 &73971181077900484 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 4826377248061997628, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
   m_PrefabInstance: {fileID: 4898914656220933880}
   m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 650081bcc202e2346b32c5edcbf16563, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!4 &8227706492431075710 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 3590177709521069958, guid: adc5b86b83d49b04a9fc99abe896f85e, type: 3}
