--- Assets/Scripts/AutomaticDoorTrigger.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 19b83b5d39ba54d429284eef91fd9101
