--- Assets/Scripts/DoorManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6fe126c229ae4ae43ad1794d7638bf5d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
