+++ Assets/Scripts/Environment/DoorScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 079a258bccb901344ade2698a14ec3fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
