--- Assets/Scripts/VentScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dcdee7ba94a3b2942a09ef44d30504c4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
